Reorganize top level of Map folder
move me
