+++ app/src/app/components/MapPage/MapPage.tsx
@@ -1,7 +1,8 @@
 'use client';
 import React, {useEffect} from 'react';
 import {MapContextMenu} from '@components/ContextMenu';
-import {DemographicMapComponent, MainMapComponent} from '@components/Map/Map';
+import {MainMap} from '@components/Map/MainMap';
+import {DemographicMap} from '@components/Map/DemographicMap';
 import SidebarComponent from '@components/sidebar/Sidebar';
 import {QueryClientProvider} from '@tanstack/react-query';
 import {queryClient} from '@utils/api/queryClient';
@@ -17,7 +18,7 @@ import {useToolbarStore} from '@/app/store/toolbarStore';
 import {useMapControlsStore} from '@/app/store/mapControlsStore';
 import {useDocumentWithSync} from '@/app/hooks/useDocumentWithSync';
 import {SaveConflictModal} from '../SaveConflictModal';
-import { migrateUserMapsFromLocalStorage } from '@/app/utils/idb/migrateUserMaps';
+import {migrateUserMapsFromLocalStorage} from '@/app/utils/idb/migrateUserMaps';
 
 interface MapPageProps {
   isEditing: boolean;
@@ -84,8 +85,8 @@ function ChildMapPage({isEditing, mapId}: MapPageProps) {
       <div className={`h-full relative w-full flex-1 flex flex-col lg:h-screen landscape:h-screen`}>
         <Topbar />
         <Flex direction="row" height="100%">
-          <MainMapComponent />
-          {showDemographicMap && <DemographicMapComponent />}
+          <MainMap />
+          {showDemographicMap && <DemographicMap />}
         </Flex>
         {toolbarLocation === 'map' && <DraggableToolbar />}
         {!!mapId && (
